Refactor: Streamline BlobHandlers moving blob operations into BlobService
