+++ Backend.WebApi/Cqrs/BlobHandlers.cs
@@ -1,10 +1,8 @@
 using Backend.Core.Cqrs.Infrastructure;
 using Backend.Core.Providers.Store;
 using Backend.DbModel.Database;
-using Backend.DbModel.Database.EntityModels;
 using Backend.WebApi.Cqrs.Infrastructure;
 using Backend.WebApi.Services;
-using Microsoft.EntityFrameworkCore;
 
 namespace Backend.WebApi.Cqrs;
 
@@ -72,38 +70,31 @@ public class BlobHandlers :
 
 	public async Task<GetBlob.Response> Handle(GetBlob query)
 	{
-		var blob = await _dbContext.Blobs
-			.Include(blob => blob.UploadedBy)
-			.Select(blob => new GetBlob.Response
-			{
-				Id = blob.Id,
-				FileName = blob.OriginalFilename,
-				FileSize = blob.FileSize,
-				PageCount = blob.PageCount,
-				UploadedAt = blob.UploadedAt,
-				UploadedByUser = blob.UploadedBy!.Fullname,
-				MimeType = blob.MimeType,
-				IsAllocated = blob.ArchiveItem != null
-			})
-			.SingleOrDefaultAsync(blob => blob.Id == query.Id);
-
+		var blob = await _blobService.GetBlobEntity(query.Id);
 		if (blob == null)
 		{
 			throw new HttpNotFoundException($"Blob with id {query.Id} not found");
 		}
 
-		return blob;
+		return new GetBlob.Response
+		{
+			Id = blob.Id,
+			FileName = blob.OriginalFilename,
+			FileSize = blob.FileSize,
+			PageCount = blob.PageCount,
+			UploadedAt = blob.UploadedAt,
+			UploadedByUser = blob.UploadedBy!.Fullname,
+			MimeType = blob.MimeType,
+			IsAllocated = blob.ArchiveItem != null
+		};
 	}
 
 
 	public async Task<IEnumerable<ListBlobs.Response>> Handle(ListBlobs query)
 	{
-		var total = await _dbContext.Blobs
-					.Where(blob => blob.ArchiveItem == null)
-					.CountAsync();
-
-		var blobs = (await _dbContext.Blobs
-			.Include(blob => blob.UploadedBy)
+		var blobEntities = await _blobService.ListBlobEntities();
+		return blobEntities
+			.OrderByDescending(blob => blob.UploadedAt)
 			.Select(blob => new ListBlobs.Response
 			{
 				Id = blob.Id,
@@ -114,33 +105,11 @@ public class BlobHandlers :
 				UploadedByUser = blob.UploadedBy!.Fullname,
 				MimeType = blob.MimeType,
 				IsAllocated = blob.ArchiveItem != null
-			})
-			.ToListAsync()) // Cannot orderBy dateTimeOffset without ToListing first
-			.OrderByDescending(blob => blob.UploadedAt)
-			.ToList();
-
-		return blobs;
+			});
 	}
 
 	public async Task Handle(DeleteBlobs command)
 	{
-		var blobs = await _dbContext.Blobs.Where(x => command.BlobIds.Contains(x.Id)).ToListAsync();
-		if (!blobs.Any())
-		{
-			throw new HttpNotFoundException("Blobs not found");
-		}
-
-		foreach (var blob in blobs)
-		{
-			var filename = blob.PathInStore.Split('/').Last();
-			var objectId = Guid.Parse(Path.GetFileNameWithoutExtension(filename));
-
-			await _objectStore.DeleteObject(objectId);
-		}
-		_dbContext.Blobs.RemoveRange(blobs);
-
-		await _dbContext.SaveChangesAsync();
-
-		await _blobService.PublishBlobsDeletedMessage(blobs);
+		await _blobService.DeleteBlobs(command.BlobIds);
 	}
 }
